Adding Final Files
This is the final version of SwoleMate. The dashboard features are still
in the works -- adding favorites and updating your workout chart -- as
are some of the matches features -- giving times and activities that
matchup for each user. However, the matches page does list the users you
match up with, it just doesn't show you the information. Otherwise, all
other features work.